Create BLFilterDocsSpans superclass.
