+++ engine/src/main/java/nl/inl/blacklab/search/lucene/BLFilterSpans.java
@@ -18,11 +18,8 @@
 package nl.inl.blacklab.search.lucene;
 
 import java.io.IOException;
-import java.util.Objects;
 
-import org.apache.lucene.search.TwoPhaseIterator;
 import org.apache.lucene.search.spans.FilterSpans;
-import org.apache.lucene.search.spans.SpanCollector;
 import org.apache.lucene.search.spans.Spans;
 
 /**
@@ -30,18 +27,13 @@ import org.apache.lucene.search.spans.Spans;
  * easily by implementing {@link #accept}
  * <p>
  * Adapted from Lucene.
- *
+ * <p>
  * NOTE: We've relaxed the visibility of some members to allow more reuse (see e.g. {@link SpansEdge}),
  * at the cost of child classes having to know more about the implementation. Specifically: match-level methods
  * are no longer final, and child classes can access atFirstInCurrentDoc and startPos, but should make sure
  * these are updated correctly when overriding methods)
  */
-public abstract class BLFilterSpans<T extends Spans> extends BLSpans {
-
-    /**
-     * The Spans object we're wrapping
-     */
-    protected final T in;
+public abstract class BLFilterSpans<T extends Spans> extends BLFilterDocsSpans<T> {
 
     /**
      * Did we advance to the first hit already through the two-phase iterator?
@@ -61,7 +53,7 @@ public abstract class BLFilterSpans<T extends Spans> extends BLSpans {
      * Wrap the given {@link T}.
      */
     protected BLFilterSpans(T in) {
-        this.in = Objects.requireNonNull(in);
+        super(in);
     }
 
     /**
@@ -69,36 +61,6 @@ public abstract class BLFilterSpans<T extends Spans> extends BLSpans {
      * NO_MORE_IN_CURRENT_DOC if iteration should move on to the next document.
      */
     protected abstract FilterSpans.AcceptStatus accept(T candidate) throws IOException;
-
-    @Override
-    public int nextDoc() throws IOException {
-        while (true) {
-            int doc = in.nextDoc();
-            if (doc == NO_MORE_DOCS) {
-                return NO_MORE_DOCS;
-            } else if (twoPhaseCurrentDocMatches()) {
-                return doc;
-            }
-        }
-    }
-
-    @Override
-    public int advance(int target) throws IOException {
-        int doc = in.advance(target);
-        while (doc != NO_MORE_DOCS) {
-            if (twoPhaseCurrentDocMatches()) {
-                break;
-            }
-            doc = in.nextDoc();
-        }
-
-        return doc;
-    }
-
-    @Override
-    public final int docID() {
-        return in.docID();
-    }
 
     @Override
     public int nextStartPosition() throws IOException {
@@ -157,73 +119,6 @@ public abstract class BLFilterSpans<T extends Spans> extends BLSpans {
                 : (startPos != NO_MORE_POSITIONS) ? in.endPosition() : NO_MORE_POSITIONS;
     }
 
-    @Override
-    public int width() {
-        return in.width();
-    }
-
-    @Override
-    public void collect(SpanCollector collector) throws IOException {
-        in.collect(collector);
-    }
-
-    @Override
-    public final long cost() {
-        return in.cost();
-    }
-
-    @Override
-    public abstract String toString();
-
-    @Override
-    public final TwoPhaseIterator asTwoPhaseIterator() {
-        TwoPhaseIterator inner = in.asTwoPhaseIterator();
-        if (inner != null) {
-            // wrapped instance has an approximation
-            return new TwoPhaseIterator(inner.approximation()) {
-                @Override
-                public boolean matches() throws IOException {
-                    return inner.matches() && twoPhaseCurrentDocMatches();
-                }
-
-                @Override
-                public float matchCost() {
-                    return inner.matchCost(); // underestimate
-                }
-
-                @Override
-                public String toString() {
-                    return "BLFilterSpans@asTwoPhaseIterator(inner=" + inner + ", in=" + in + ")";
-                }
-            };
-        } else {
-            // wrapped instance has no approximation, but 
-            // we can still defer matching until absolutely needed.
-            return new TwoPhaseIterator(in) {
-                @Override
-                public boolean matches() throws IOException {
-                    return twoPhaseCurrentDocMatches();
-                }
-
-                @Override
-                public float matchCost() {
-                    return in.positionsCost(); // overestimate
-                }
-
-                @Override
-                public String toString() {
-                    return "BLFilterSpans@asTwoPhaseIterator(in=" + in + ")";
-                }
-            };
-        }
-    }
-
-    @Override
-    public float positionsCost() {
-        //throw new UnsupportedOperationException(); // asTwoPhaseIterator never returns null
-        return in.positionsCost();
-    }
-
     /**
      * Returns true if the current document matches.
      * <p>
@@ -232,13 +127,6 @@ public abstract class BLFilterSpans<T extends Spans> extends BLSpans {
     @SuppressWarnings("fallthrough")
     protected boolean twoPhaseCurrentDocMatches() throws IOException {
         atFirstInCurrentDoc = false;
-
-        // FIXME: this seems to cause TestSearchHitGroups.testHitGroups() to fail
-        //        with fast path and filter. The reason seems to be that in this case,
-        //        we use the SpanQuery to find all matching docs. But this operation
-        //        doesn't call nextDoc() on the Spans object; it only uses the two-phase
-        //        iterator and this matches() method tries to advance the Spans object,
-        //        but that hasn't been initialized yet (currentDoc is still at -1).
         startPos = in.nextStartPosition();
         assert startPos != NO_MORE_POSITIONS;
         for (;;) {
@@ -258,17 +146,4 @@ public abstract class BLFilterSpans<T extends Spans> extends BLSpans {
             }
         }
     }
-
-    @Override
-    protected void passHitQueryContextToClauses(HitQueryContext context) {
-        if (in instanceof BLSpans)
-            ((BLSpans) in).setHitQueryContext(context);
-    }
-
-    @Override
-    public void getMatchInfo(MatchInfo[] relationInfo) {
-        if (childClausesCaptureMatchInfo && in instanceof BLSpans)
-            ((BLSpans) in).getMatchInfo(relationInfo);
-    }
-
 }
